refactor: Operator를 Model group으로 이동 #4
