+++ Calculator/CalculatorTests/FormulaTests.swift
@@ -13,45 +13,6 @@ extension String {
     }
 }
 
-enum Operator: Character, CalculateItem, CaseIterable {
-    case add = "+"
-    case subtract = "-"
-    case divide = "/"
-    case multiply = "*"
-    
-    func calculate(lhs: Double, rhs: Double) -> Double {
-        switch self {
-        case .add:
-            return add(lhs: lhs, rhs: rhs)
-        case .divide:
-            return didvde(lhs: lhs, rhs: rhs)
-        case .multiply:
-            return multiply(lhs: lhs, rhs: rhs)
-        case .subtract:
-            return subtract(lhs: lhs, rhs: rhs)
-        }
-    }
-    
-    private func add(lhs: Double, rhs: Double) -> Double {
-        lhs + rhs
-    }
-    
-    private func subtract(lhs: Double, rhs: Double) -> Double {
-        lhs - rhs
-    }
-    
-    private func didvde(lhs: Double, rhs: Double) -> Double {
-        if rhs == .zero {
-            return .nan
-        }
-        return lhs / rhs
-    }
-    
-    private func multiply(lhs: Double, rhs: Double) -> Double {
-        lhs * rhs
-    }
-}
-
 struct Formula {
     var operands = CalculatorItemQueue<Double>()
     var operators = CalculatorItemQueue<Character>()
